Add basic zip64 support so we can handle zips larger than 4GB

+++ Wabbajack.Compression.Zip/ExtractedEntry.cs
@@ -5,6 +5,6 @@ public class ExtractedEntry
     public long FileOffset { get; set; }
     public string FileName { get; set; }
     public long CompressedSize { get; set; }
-    public uint UncompressedSize { get; set; }
+    public long UncompressedSize { get; set; }
     public short CompressionMethod { get; set; }
 }
